Modularize file_utils
